+++ packages/web/src/components/ui/dropdown-menu.tsx
@@ -105,9 +105,9 @@ const DropdownMenuCheckboxItem = React.forwardRef<
     checked={checked}
     {...props}
   >
-    <span className="absolute left-2 flex h-3.5 w-3.5 items-center justify-center">
-      <DropdownMenuPrimitive.ItemIndicator>
-        <Check className="h-4 w-4" />
+    <span className="absolute left-2 flex h-4 w-4 items-center justify-center rounded-sm border border-muted bg-transparent">
+      <DropdownMenuPrimitive.ItemIndicator className="flex items-center justify-center text-current">
+        <Check className="h-3.5 w-3.5" />
       </DropdownMenuPrimitive.ItemIndicator>
     </span>
     {children}