+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Changed the default `/repos` pagination size to 20. [#706](https://github.com/sourcebot-dev/sourcebot/pull/706)
+- Added checkbox to exclude archived and forked repositories from settings dropdown in the query search tab. [#663](https://github.com/sourcebot-dev/sourcebot/issues/663)
 
 ## [4.10.7] - 2025-12-29